+++ src/main/java/com/meti/app/HostALocalServer.java
@@ -2,9 +2,9 @@ package com.meti.app;
 
 import com.meti.lib.fx.Controller;
 import com.meti.lib.fx.ControllerLoader;
-import com.meti.lib.net.Client;
-import com.meti.lib.net.Server;
-import com.meti.lib.net.SocketConnection;
+import com.meti.lib.net.client.Client;
+import com.meti.lib.net.server.Server;
+import com.meti.lib.net.server.SocketConnection;
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
 import javafx.scene.control.TextField;
