+++ src/main/java/com/meti/app/ServerDisplay.java
@@ -3,9 +3,9 @@ package com.meti.app;
 import com.meti.lib.util.BufferedConsole;
 import com.meti.lib.fx.Controller;
 import com.meti.lib.fx.PostInitializable;
-import com.meti.lib.net.Client;
-import com.meti.lib.net.Server;
-import com.meti.lib.net.SocketConnection;
+import com.meti.lib.net.client.Client;
+import com.meti.lib.net.server.Server;
+import com.meti.lib.net.server.SocketConnection;
 import com.meti.lib.util.Finalizable;
 import javafx.animation.AnimationTimer;
 import javafx.application.Platform;
